+++ firebase-dataconnect/CHANGELOG.md
@@ -8,6 +8,8 @@
   [#7720](https://github.com/firebase/firebase-android-sdk/pull/7720))
 - [changed] Wired up implementation for serving query responses from cache.
   [#7759](https://github.com/firebase/firebase-android-sdk/pull/7759))
+- [changed] DataSource wired up and entity paths extracted from response.
+  [#NNNN](https://github.com/firebase/firebase-android-sdk/pull/NNNN))
 
 # 17.1.3